third_party/blink/web_tests/external/wpt/webrtc/RTCIceCandidate-constructor-expected.txt

This is a testharness.js-based test.
PASS new RTCIceCandidate()
PASS new RTCIceCandidate({})
PASS new RTCIceCandidate({ ... }) with manually filled default values
PASS new RTCIceCandidate({ sdpMid: null, sdpMLineIndex: null })
PASS new RTCIceCandidate({ candidate: '' })
PASS new RTCIceCandidate({ candidate: null })
PASS new RTCIceCandidate({ ... }) with valid candidate string only
FAIL new RTCIceCandidate({ sdpMid: 'audio' }) Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty.
FAIL new RTCIceCandidate({ sdpMLineIndex: 0 }) Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty.
FAIL new RTCIceCandidate({ sdpMid: 'audio', sdpMLineIndex: 0 }) Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty.
FAIL new RTCIceCandidate({ candidate: '', sdpMid: 'audio' } Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty.
FAIL new RTCIceCandidate({ candidate: '', sdpMLineIndex: 0 } Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty.
FAIL new RTCIceCandidate({ ... }) with valid candidate string and sdpMid assert_equals: sdpMLineIndex expected (object) null but got (number) 0
FAIL new RTCIceCandidate({ ... }) with invalid candidate string and sdpMid assert_equals: sdpMLineIndex expected (object) null but got (number) 0
FAIL new RTCIceCandidate({ ... }) with non default value for all fields assert_equals: usernameFragment expected (string) "test" but got (undefined) undefined
FAIL new RTCIceCandidate({ ... }) with invalid sdpMid Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty.
FAIL new RTCIceCandidate({ ... }) with invalid sdpMLineIndex Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty.
Harness: the test ran to completion.
